+++ src/main/java/lotto/domain/LottoRule.java
@@ -1,4 +1,4 @@
-package lotto.config;
+package lotto.domain;
 
 public enum LottoRule {
